feat: Add .gitattributes and update .gitconfig and setup.sh
This commit introduces a .gitattributes file to handle Jupyter notebook
files (*.ipynb) with a custom filter 'clean_ipynb'. This filter cleans
the notebook by removing outputs and execution counts.

The .gitconfig file is updated to include the new .gitattributes file
and to define the 'clean_ipynb' filter.

The setup.sh script is also updated to include the .gitattributes file
in the list of dotfiles to be handled by the script.