feat: Add module reload function and update plugins
This commit introduces a new function `_G.reload_module(name)` in the
`init.lua` file. This function allows for reloading a module by setting
its entry in `package.loaded` to `nil` and then requiring it again.

The `highlight_undo` plugin has been commented out in the
`highlight_undo.lua` file. This change was made because the plugin was
causing issues with the undo and redo functionality.

The `hlslens` plugin has been updated in the `hlslens.lua` file to
include the `lazy = false` configuration. This change was made to
improve the performance of the plugin.

The `luminate` plugin has been updated in the `mei.lua` file. The
`event` configuration has been replaced with `keys` configuration and
the setup function has been simplified. This change was made to improve
the performance of the plugin and to make it more customizable.